Kmesh: fix node clean failed in rb tree
When deleteing a red-black tree, do not use the rb_erase interface.
This inerface will cause the red-black tree to be rebalanced. As a
result, nodes are lost during deletion. The correct approach would
be to use the subsequent output node and directly release and leave
the root node pointer empty.

Signed-off-by: bitcoffee <liuxin350@huawei.com>