Kmesh: Introduce a new method to query whether a pod is managed by Kmesh
It is difficult to obtain whether the current pod is managed by Kmesh
in the ebpf sockops. Now, a new general solution is available after
version 5.14 by getsockops. However, in the system before version 5.14,
a new solutions by bpf map is required. That's what CNI does.

A map is introduced and pinned to a fixed path when the ebpf program is
loaded. The CNI inserts the allocated IP address into the map so that
the ebpf program can determine whether the current IP address is managed
by the Kmesh. If a pod exits, the corresponding ip address hould be
cleaned out of the map.

Signed-off-by: bitcoffee <liuxin350@huawei.com>